+++ examples/filter.py
@@ -1,0 +1,14 @@
+from typing import Iterator
+
+
+def func(x: int):
+    """Function."""
+
+
+def gen() -> Iterator[int]:
+    """Generator."""
+    yield 1
+
+
+class C:
+    """Class."""