+++ mkapi/core/page.py
@@ -1,7 +1,7 @@
 """This module provides a Page class that works with other converter."""
 import re
 from dataclasses import InitVar, dataclass, field
-from typing import Iterator, List, Union
+from typing import Iterator, List, Tuple, Union
 
 from mkapi import utils
 from mkapi.core import postprocess
@@ -30,8 +30,12 @@ class Page:
     source: InitVar[str]
     abs_src_path: str
     abs_api_paths: List[str] = field(default_factory=list, repr=False)
+    filters: List[str] = field(default_factory=list, repr=False)
     markdown: str = field(init=False, repr=False)
     nodes: List[Union[Node, Code]] = field(default_factory=list, init=False, repr=False)
+    headings: List[Tuple[int, str]] = field(
+        default_factory=list, init=False, repr=False
+    )
 
     def __post_init__(self, source):
         self.markdown = "\n\n".join(self.split(source))
@@ -47,15 +51,21 @@ class Page:
     def split(self, source: str) -> Iterator[str]:
         cursor = 0
         callback = self.resolve_link_from_base
-        for index, match in enumerate(MKAPI_PATTERN.finditer(source)):
+        index = 0
+        for match in MKAPI_PATTERN.finditer(source):
             start, end = match.start(), match.end()
             if cursor < start:
                 markdown = source[cursor:start].strip()
                 if markdown:
                     yield self.resolve_link(markdown)
+            cursor = end
             heading, name = match.groups()
             level = len(heading)
             name, filters = utils.split_filters(name)
+            if not name:
+                self.filters = filters
+                continue
+            filters = utils.update_filters(self.filters, filters)
             if "code" in filters:
                 code = get_code(name)
                 self.nodes.append(code)
@@ -66,8 +76,10 @@ class Page:
                 postprocess.transform(node, filters)
                 self.nodes.append(node)
                 markdown = node.get_markdown(level, callback=callback)
+                if level:
+                    self.headings.append((level, node.object.id))
             yield node_markdown(index, markdown, filters)
-            cursor = end
+            index += 1
         if cursor < len(source):
             markdown = source[cursor:].strip()
             if markdown: